test(changelog-github): move template helper tests to render-template.test.ts
